awdb_event_logs/requirements.txt
pandas
sqlalchemy
Flask-SQLAlchemy
flask-restx
pymssql
python-dotenv